Tweaked test specification.
Originally I had intended to remove any extraneous parentheses such as
those that surround language expressions. Unfortunately, this proved to be
more difficult than I had anticipated as it is hard to differentiate said
extraneous parentheses from meaningful ones such as those used to invoke a
method. It is for these reasons that I have changed the test specification
to allow extraneous parentheses.